Don't Hate Me Update #2
Removed all dependencies on the CommonServiceLocator.
-Added a custom IServiceLocator interface with a smaller API.
-Added additional extension methods for the locator.
-Added an IBuilder interface with a BuildUp method.
-Composed IServiceLocator, IRegistry and IBuilder into IContainer.
-Added a static gateway called IoC with the same API as Caliburn.Micro. This replaces the ServiceLocator.Current functionality.
-Updated all container adapters.

Removed Prism Integration
Prism has a dependency on CSL which we now do not have. One of the original reasons for our integration was to enable other containers on Prism. Prism has solved this issue themselves recently, so there's no need for us to do it. Furthermore, I feel that there's not much need to use Prism in addition to Caliburn if you are using Caliburn correctly and it was just another dependency in the project and a maintenance issue for me.. If you do desire to use them both together, you have at least two options:
1. Get the old code out of the repo and modify it for your uses.
2. Use Caliburn.Micro with Prism, which will probably be a much nicer combination.

FYI, tomorrow I will probably remove CaliburnApplication and replace it with a Bootstrapper. So, you may want to wait to update until after that...